Rename objects to datahandling, move loading and mixins to datahandling
