+++ eitprocessing/datahandling/loading/__init__.py
@@ -1,9 +1,10 @@
 from functools import reduce
 from pathlib import Path
 
-from eitprocessing.objects.datacollection import DataCollection
-from eitprocessing.objects.eitdata import EITData, Vendor
-from eitprocessing.objects.sequence import Sequence
+from eitprocessing.datahandling.datacollection import DataCollection
+from eitprocessing.datahandling.eitdata import EITData, Vendor
+from eitprocessing.datahandling.loading import draeger, sentec
+from eitprocessing.datahandling.sequence import Sequence
 
 
 def load_eit_data(
@@ -51,7 +52,7 @@ def load_eit_data(
     pixel_impedance = sequence.eit_data["raw"].pixel_impedance
     ```
     """
-    from eitprocessing.loading import draeger, sentec, timpel  # not in top level to avoid circular import
+    from eitprocessing.datahandling.loading import timpel  # not in top level to avoid circular import
 
     vendor = _ensure_vendor(vendor)
     load_func = {
