+++ eitprocessing/datahandling/loading/timpel.py
@@ -6,12 +6,12 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
-from eitprocessing.loading import load_eit_data
-from eitprocessing.objects.continuousdata import ContinuousData
-from eitprocessing.objects.datacollection import DataCollection
-from eitprocessing.objects.eitdata import EITData, Vendor
-from eitprocessing.objects.phases import MaxValue, MinValue, QRSMark
-from eitprocessing.objects.sparsedata import SparseData
+from eitprocessing.datahandling.continuousdata import ContinuousData
+from eitprocessing.datahandling.datacollection import DataCollection
+from eitprocessing.datahandling.eitdata import EITData, Vendor
+from eitprocessing.datahandling.loading import load_eit_data
+from eitprocessing.datahandling.phases import MaxValue, MinValue, QRSMark
+from eitprocessing.datahandling.sparsedata import SparseData
 
 if TYPE_CHECKING:
     from pathlib import Path
