+++ eitprocessing/datahandling/mixins/equality.py
