+++ eitprocessing/datahandling/mixins/slicing.py
