+++ eitprocessing/datahandling/sparsedata.py
